union: No longer perform stat of "/" when mounting filesystems.
Doing so is an expensive operation on some virtual filesystems. Skip
it, make it a responsibility of the caller to ensure that provided
filesystems to be mounted are okay.

As a result, we no longer have a valid mod time for root directory, but
that's a fair sacrifice since it can be expensive to compute, yet is
not used much.