vfsutil: Remove x/tools/godoc/vfs dependency.
Importing the golang.org/x/tools/godoc/vfs is not justified and easily
avoidable, since this vfsutil package deals with http.FileSystem
interface, not godoc/vfs.FileSystem interface.

openStat is already returning what http.FileSystem.Open returns, which
is http.File. We can use that.

This simplifies the code, and reduces the number of imported packages.

Resolves #6.